refactor: muda script de download de mapas para serviço
